added the feature to change usernames
Changed the user variable to be an object instead of just username, now it containts username,userId and profilePicture, now it's being stored in the database as the user object and not just usernames.

Instead of searching a user with their usernames now we are searching them with their userIds.