fixed user not get deleted from typing state when other user continues typing
The stopped-typing-to-client function wasn't working as planned, because the typing-to-client function was working with the previous version of the state and not the one that was mutated by stopped-typing socket emit function. Fixed this by changing the way the state gets updated, now state gets set by the functional way.