added security when joining rooms in backend
Made it so that people won't be able to type each other's userId's and join their private rooms, when joining a room the backend checks if the room nameincludes client's userId if so, joins the room.

Also, deleted the maximum length requirement for room names in backend, implemented another way of security instead. Now it only exists in the front end.